Improve diagnostics on header/impl classes when scopes don't match
Try to report most of the errors on the actual members of the impl
class. In many cases, there's a 1:1 mapping of header to impl class
members, so the error "some members are not implemented" on the class
declaration itself is redundant. Exceptions include functions/properties
from supertypes (there may be no other place to report an error in
signature mismatch in this case), functions/properties not marked with
'impl' and default constructors (the checker simply is not run for them)

 #KT-18447 Fixed